test: utility is divided from board_test
